requirements.txt
RPi.GPIO==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2
